CORE-1432 Refactor EditTool's formatting functions into new file
Refactored EditTool's `mapPropsToValues` and `formatSubmission` into a
new `formatters.js` file.
